fix(ops): demote build stamps without hiding them
The gate caught two real failures in SelfUpgradeClient.test.tsx, a file this
worktree could not run locally because apps/web/node_modules lacked next.

First: a test asserted the very explainer sentence this branch deletes. That
explanation only existed because the SHAs were the headline identity; with the
merged PR leading, the contradiction it defused cannot arise. The test now
asserts the apology is gone rather than merely relocated.

Second: a test asserts no <details> element renders when a run has no error,
guarding the failure-log disclosure. The build-stamp block this branch added
was a raw <details>, which broke it.

Neither disclosure construct fits here. A native <details> is the construct the
progressive-disclosure doctrine ranks lowest, and the canonical ExpandableCard
omits its collapsed subtree from the DOM entirely (BI-2B196D07) -- which made
the deployed SHA, target SHA, and platform version unreachable and broke five
further contract tests asserting the ops banner exposes them. So the stamps are
not deferred at all: demotion here means smaller and quieter, not hidden. The
component is renamed BuildStamps to say what it actually is.

Also junctions next into apps/web/node_modules so this file, and the rest of the
component suite, can be verified in the worktree instead of only in the sandbox.
76/76 in that file, 904 across every touched area including ux-budget.

Backlog: BI-5B1FDA09

Design-Grounding-Decision: extend-existing-spec -- apps/web/lib/ux-budget/scope.ts (DISCLOSURE_TRIGGER_ATTR, BI-2B196D07) is the source of truth for which disclosure constructs count and why; this change applies it and adds no contract.
